+++ tests/ArrayAdapterTest.php
@@ -17,6 +17,7 @@ class ArrayAdapterTest extends TestCase
 {
 
     /** @test
+     * @throws \Exception
      */
     public function stores_and_retrieves_data() : void
     {
@@ -41,15 +42,16 @@ class ArrayAdapterTest extends TestCase
      */
     public function stores_and_retrieves_response() : void
     {
+        $responseBody = 'test';
         $request = new Request('GET', 'www.test.de');
-        $response = new Response(200, [], null, '1337');
+        $response = new Response(200, [], $responseBody);
 
         $arrayAdapter = new ArrayAdapter();
         $arrayAdapter->saveResponse($request, $response);
 
         $storedResponse = $arrayAdapter->getResponse($request);
 
-        static::assertEquals($response, $storedResponse);
+        static::assertEquals($responseBody, (string) $storedResponse->getBody());
     }
 
     /** @test
@@ -58,11 +60,53 @@ class ArrayAdapterTest extends TestCase
     public function stored_value_gets_invalidated_when_expired() : void
     {
         $request = new Request('GET', 'www.test.com');
-        $response = new Response(200, [], null, '1337');
+        $response = new Response(200, [], 'test');
 
         $arrayAdapter = new ArrayAdapter(new Repository(['cache' => ['ttl' => 0]]));
         $arrayAdapter->saveResponse($request, $response);
 
         static::assertNull($arrayAdapter->getResponse($request));
+    }
+
+    /**
+     * @test
+     */
+    public function does_not_store_empty_values() : void
+    {
+        $request = new Request('GET', 'www.test.com');
+        $nullResponse = new Response(200, [], null);
+
+        $arrayAdapter = new ArrayAdapter();
+        $arrayAdapter->saveResponse($request, $nullResponse);
+
+        static::assertNull($arrayAdapter->getResponse($request));
+
+        $emptyResponse = new Response(200, [], '');
+
+        $arrayAdapter = new ArrayAdapter();
+        $arrayAdapter->saveResponse($request, $emptyResponse);
+
+        static::assertNull($arrayAdapter->getResponse($request));
+    }
+
+    /**
+     * @test
+     */
+    public function stores_empty_values_when_allowed() : void
+    {
+        $request = new Request('GET', 'www.test.com');
+        $nullResponse = new Response(200, [], null);
+
+        $arrayAdapter = new ArrayAdapter(new Repository(['cache' => ['allow_empty' => true]]));
+        $arrayAdapter->saveResponse($request, $nullResponse);
+
+        static::assertEmpty((string) $arrayAdapter->getResponse($request)->getBody());
+
+        $emptyResponse = new Response(200, [], '');
+
+        $arrayAdapter = new ArrayAdapter(new Repository(['cache' => ['allow_empty' => true]]));
+        $arrayAdapter->saveResponse($request, $emptyResponse);
+
+        static::assertEmpty((string) $arrayAdapter->getResponse($request)->getBody());
     }
 }