+++ tests/config/app.php
@@ -2,10 +2,18 @@
 
 return [
     'cache' => [
-        'driver'  => 'file',
-        'options' => [
-            'path' => 'cache'
+        'driver'      => 'redis',
+        'options'     => [
+            'database' => [
+                'cluster' => false,
+                'default' => [
+                    'host'     => '127.0.0.1',
+                    'port'     => 6379,
+                    'database' => 0,
+                ],
+            ]
         ],
-        'ttl'     => 900
+        'ttl'         => 900,
+        'allow_empty' => false
     ]
 ];